build: script.zsh: Introduce pedantic quoting
From what I can tell, Zsh offers by default a much saner approach
to word splitting when performing parameter substitution; however,
it's probably safest to quote parameters as though Zsh were as
insane as other shells.

Furthermore, quotes have been removed from places where they are
not needed, and parameter-expansion braces have been removed unless
they are necessary for proper parsing. `((...))' and `[[...]]'
constructs have been used where appropriate, instead of the
old `[...]' construct.

One aspect of quoting that is still not handled well is the
programmatic construction of the `.depend' makefile; a theoretically
necessary task is not being performed: File names need to be quoted
or otherwise escaped to ensure that both `make' and the shell that
`make' invokes are receiving the correct data. This is a somewhat
complicated issue that is better left to a time when the code in
question has been simplified.