build: Move non-deployment code from `script.zsh' to `build.sh'
At this point, `build.sh' is just a superset of `script.zsh';
in particular, `build.sh' defines some functions and then
includes `script.zsh'.

The deployment system should be able to continue using `script.zsh'
just as it always has.

One point of contention is the exact nature of the `update_bespin'
function, which is used both as a build-time and deployment-time
function. This is why `build.sh' has not been separated completely
from `script.zsh', and this is why the target `bespin' of `Makefile'
runs `script.zsh' rather than `build.sh'.