build: build.sh: Move functions outside of the `depend' function
There were a bunch of functions defined within the body of the `depend'
function; this commit moves their definitions out of the body, fixes
indentation, and replaces unnecessary tabs with spaces.

Also, many of the rule lines produced for `.depend' ended with white
space; this has been corrected by appending text to the prerequisites
only when necessary.