build: build.sh: Replace literal tab character with something more deliberate
A makefile is often sensitive to the tab character being used according to
strict rules; in constructing the `.depend' makefile, the code was using
a literal tab character, but it seems dangerous to do so, because it might
be easily replaced with spaces by mistake.

This commit obviates the problem by specifying more deliberately that a tab
character should be used.