build: Explicitly separate build-code from deploy-code (bespin-code)
As this source code is meant to build a website, it's not surprising
that it's tailored to deploying the results to a particular server
(the server seems to be referred to as `bespin').

While this is very practical, it's not actually necessary for the
results to be deployed to that particular server; someone might
want to put up a mirror on a different server.

This commit makes an initial attempt at separating the building
of this website from the deployment to a particular server; what
this means is that makefile code for doing the deployment has been
segregated in its own file:

  bespin.mk

and the link between `script.zsh' and `build.sh' has been severed
by copying code as necessary from `script.zsh' to `build.sh'; the
duplication of code is unfortunate, but necessary, especially
because it is desirable to keep backwards compatibility with the
existing infrastructure, which expects a self-contained `script.zsh'.