cleanup: Remove accidentally commited `*.html' products
Running `make clean' revealed that some of the `*.html' files
that are meant to be generated from the corresponding `*.scr'
files have been accidentally committed as tracked files.

This commit relieves `git' of the burden of tracking those files
by simply committing the results of `make clean'; this commit
also simplifies the `.gitignore' file.