build: build.sh: Use a recursive glob to find the dependencies
I noticed that `./build.sh clean' has been using a recursive glob
to find all possibly relevant `*.scr' files; in contrast, the
`./build.sh depend' command does not use a recursive glob.

As it turns out, this discrepancy is purposeful; not every `*.scr'
file is meant to be compiled into a corresponding `*.html' file.
In particular, it would seem the only top-level `*.scr' file that
is meant to be so processed is `index.scr'; the other top-level
`*.scr' files are ancillary.

Put another way: Other than `index.scr', every `*.scr' file that is
meant to be compiled into a corresponding `*.html' file is located
in a subdirectory; furthermore, each such subdirectory does not
have any subdirectory of its own.

However, one day, such a subdirectory might have a subdirectory
of its own; this commit takes that possibility into account, and
therefore introduces a recursive glob within each subdirectory.

Currently, using a recursive glob changes nothing in practice, but
it might be useful to handle this more general case so as to be
prepared for the future.