build: build.sh: compat: Comment out `local' declarations
One of the most widely implemented non-POSIX features is the `local'
built-in, which declares a variable to be dynamically local to a
function; it is essential for producing complicated scripts where
one part is prevented from clobbering another part.

Alas, POSIX does not yet specify `local', so these declarations
have been commented out, waiting for the day when they'll be
considered acceptable. For now, they simply document what the
scope of the variables in question should be.